Dumping all changes, because  I did too much without committing.
Fixed bug that child Octrees would be spawning farther away (turns out it was [1, 2, 4] not [1, 1, 1] for the cells). Disabled StaticOctree class. Added debug rendering for boxes. Used mainly for the Octree. Added a "program" or scene for the Octree. Imported Assimp & other GL stuff, like the Camera and Shader. Refactored a lot. Additionally: const-correctness in Shape classes. Made shapes not solely a component, which forced a Transform (kinda) onto the Shapes themselves. Will figure that out particular part later. Renamed the movement class from SimplePhysicsBehavior to OctreeMovement. Added several shaders, one for which debug rendering of lines, which is how Boxes are drawn. Updated namespaces