Added sphere visuals, fixed physics, added collision resolving
Firstly, the rendering was slightly off. (by 2, because dividing or multiplying by two is hard), made several classes complete types, so that they could be as concrete objects, and bounds, but without a position.
Changed the movement component so that it does not go outside its assigned bounds, in this case the Octree (cage). Fixed physical inaccuracies for Octree leaf positions which were caused by my inability differentiate the size and half-size of a box. Go me! :)